pyproject.toml
[project]
name = "e-commerce-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "google-auth>=2.48.0",
    "google-cloud-pubsub>=2.35.0",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
]
